requirements.txt
pytest==7.1.2
requests==2.27.1
PyYAML==6.0
jinja2==3.1.2
cookiecutter
